+++ srcs/app_server/pong_online/static/pong_online/lobby.js
@@ -1,72 +1,54 @@
-import {fetch_html_replace_dynamicDIV_activate_js, fetch_with_internal_js} from "../userManagement/land.js";
+import {router, ws} from "../userManagement/land.js"
 
-class js_wrapper {
-	activate() {
-		throw new Error("Subclasses must override this method.");
-	}
+class Lobby extends HTMLElement {
+    constructor() {
+        super();
 
-	deactivate () {
-		throw new Error("Subclasses must override this method.");
-	}
-}
+		this.username = this.getAttribute('data-username');
+		console.log("username lobby:", this.username);
 
-	
-class Lobby extends js_wrapper {
-	static counter = 0;
+		ws.send(JSON.stringify({type: 'username', 'username': this.username}));
 
-	constructor(ws, username) {
-		super();
-		Lobby.counter++;
-		console.log("lobby instances: ", Lobby.counter);
-        this.ws = ws;
-		this.username = username;
-		console.log("username:", username);
-		
-		this.modus = null;
-    }
-	
-	send_initial_data_to_server(e) {
-		console.log('WebSocket connection established');
-		this.ws.send(JSON.stringify({type: 'username', 'username': this.username}));
-		this.ws.send(JSON.stringify({type: 'lobby_update', 'action': 'display'}));
-	}
-	
-	activate() {
-		this.ws.onmessage = (e) => this.handle_message(e);
-		this.create_game_button = document.getElementById('create_game');
+
+        this.innerHTML = /*html*/`
+			<div>
+				<button class="btn buttonblue" id="create_game">Create remote game</button>
+				<button class="btn buttonblue" id="create_tournament">Create tournament</button>
+				<button class="btn buttonblue" id="play_ai">Play against Ai</button>
+				<button class="btn buttonblue" id="play_local">Play local</button>
+			</div>
+			
+			<div id="remote_game_list"></div>
+			<div id="tournament_list"></div>
+        `;
+
+		ws.onmessage = (e) => this.handle_message(e);
+
+		this.create_game_button = this.querySelector('#create_game');
 		this.create_game_button.addEventListener('click', 
 			this.handle_create_game_button_click);
 		
-		this.create_tournament_button = document.getElementById('create_tournament');
+		this.create_tournament_button = this.querySelector('#create_tournament');
 		this.create_tournament_button.addEventListener('click',
 			this.handle_create_tournament_button_click);
 		
-		this.play_local_button = document.getElementById('play_local');
+		this.play_local_button = this.querySelector('#play_local');
 		this.play_local_button.addEventListener('click', 
 			this.handle_play_local_button_click);
 		
-		this.play_ai_button = document.getElementById('play_ai');
+		this.play_ai_button = this.querySelector('#play_ai');
 		this.play_ai_button.addEventListener('click',
 			this.handle_play_ai_button_click);
 		
-		this.remote_game_list_DIV = document.getElementById('remote_game_list');
-		this.tournament_list_DIV = document.getElementById('tournament_list');
+		this.remote_game_list_DIV = document.querySelector('#remote_game_list');
+		this.tournament_list_DIV = document.querySelector('#tournament_list');
 
 		// this.ws.onopen = (e) => this.send_initial_data_to_server(e);
 
-		this.ws.send(JSON.stringify({type: 'lobby_update', 'action': 'display'}));
-	}
+		ws.send(JSON.stringify({type: 'lobby_update', 'action': 'display'}));
+    }
 
-	deactivate() {
-		this.create_game_button.removeEventListener('click', this.handle_create_game_button_click);
-        this.create_tournament_button.removeEventListener('click', this.handle_create_tournament_button_click);
-        this.play_local_button.removeEventListener('click', this.handle_play_local_button_click);
-        this.play_ai_button.removeEventListener('click', this.handle_play_ai_button_click);
-
-        this.ws.onmessage = null;
-	}
-
-    handle_message(e) {
+	handle_message(e) {
 		console.log('handle message');
 		try {
 			const data = JSON.parse(e.data);
@@ -91,32 +73,32 @@ class Lobby extends js_wrapper {
 				console.log('Error parsing JSON:', error);
 		}
 	}
-				
-		handle_create_game_button_click = () => {
-			console.log("Create game button clicked");
-			this.ws.send(JSON.stringify({type: 'lobby_update', 'action': 'create',
-			'username': this.username, 'modus': 'remote'}));
-		}
-		
-		handle_create_tournament_button_click = () => {
-			console.log("Create tournament button clicked");
-			this.ws.send(JSON.stringify({type: 'lobby_update', 'action': 'create_tournament',
-			'username': this.username}));
-		}
-		
-		handle_play_local_button_click = () => {
-			console.log("Local game button clicked");
-			this.ws.send(JSON.stringify({type: 'lobby_update', 'action': 'join', 
-			'username': this.username, 'modus': 'local'}))
-		}
-		
-		handle_play_ai_button_click = () => {
-			console.log("AI Opponent button clicked");
-			this.ws.send(JSON.stringify({type: 'lobby_update', 'action': 'join',
-			'username': this.username, 'modus': 'ai'}));
-		}
-		
-		update_lobby(data) {
+
+	handle_create_game_button_click = () => {
+		console.log("Create game button clicked");
+		ws.send(JSON.stringify({type: 'lobby_update', 'action': 'create',
+		'username': this.username, 'modus': 'remote'}));
+	}
+	
+	handle_create_tournament_button_click = () => {
+		console.log("Create tournament button clicked");
+		ws.send(JSON.stringify({type: 'lobby_update', 'action': 'create_tournament',
+		'username': this.username}));
+	}
+	
+	handle_play_local_button_click = () => {
+		console.log("Local game button clicked");
+		ws.send(JSON.stringify({type: 'lobby_update', 'action': 'join', 
+		'username': this.username, 'modus': 'local'}))
+	}
+	
+	handle_play_ai_button_click = () => {
+		console.log("AI Opponent button clicked");
+		ws.send(JSON.stringify({type: 'lobby_update', 'action': 'join',
+		'username': this.username, 'modus': 'ai'}));
+	}
+
+	update_lobby(data) {
 		let matches_info = data.matches_info;
 		let tournaments_info = data.tournaments_info;
 		console.log("lobby_update tm-info: ", tournaments_info);
@@ -141,7 +123,7 @@ class Lobby extends js_wrapper {
 				let joinButton = matchElement.querySelector('.join');
 				joinButton.addEventListener('click', () => {
 					console.log("Join button clicked for match id: ", id);
-					this.ws.send(JSON.stringify({type: 'lobby_update', 'action': 'join',
+					ws.send(JSON.stringify({type: 'lobby_update', 'action': 'join',
 					'match_id': id, 'username': this.username, 'modus': 'remote'}));
 				});
 			})(match_id);
@@ -163,7 +145,7 @@ class Lobby extends js_wrapper {
 				let joinButton = tournamentElement.querySelector('.join');
 				joinButton.addEventListener('click', () => {
 					console.log("Join button clicked for tournament id: ", id);
-					this.ws.send(JSON.stringify({type: 'lobby_update', 'action': 'join_tournament',
+					ws.send(JSON.stringify({type: 'lobby_update', 'action': 'join_tournament',
 					'tournament_id': id, 'username': this.username}));
 				});
 			})(tournament_id);
@@ -171,58 +153,29 @@ class Lobby extends js_wrapper {
 	}
 
 	join_game(modus) {
-		this.ws.modus = modus;
-		fetch_html_replace_dynamicDIV_activate_js('/pong_online', true, () => {
-			this.ws.send(JSON.stringify({type: 'get_game_data'}));
-		});
+		// this.ws.modus = modus;
+		// fetch_html_replace_dynamicDIV_activate_js('/pong_online', true, () => {
+		history.pushState("", "", "/pong_online");
+		router();
+		ws.send(JSON.stringify({type: 'get_game_data'}));
+		// });
+		console.log("join game");
 	}
 
 	join_tournament(server_data) {
-		fetch_html_replace_dynamicDIV_activate_js('/tournament/' + server_data.tournament_id + '/', true, () => {
-			this.ws.send(JSON.stringify({type: 'tournament_lobby_update', 'tournament_id': server_data.tournament_id}));
-		});
-
-		// document.body.innerHTML = '';
-		// fetch('/tournament/' + server_data.tournament_id + '/')
-		// 	.then(response => response.text())
-		// 	/* .then(data => {
-		// 		document.body.innerHTML = data
-		// 	}) */
-		// 	.then(fetch_data => {
-		// 		console.log('/tournament/' + server_data.tournament_id)
-		// 		// Create a temporary DOM div element
-		// 		var tempDiv = document.createElement('div');
-
-		// 		// Set its innerHTML to the fetched HTML data
-		// 		tempDiv.innerHTML = fetch_data;
-		
-		// 		// Extract the src attribute from the script tag
-		// 		var scriptSrc = tempDiv.querySelector('script').src;
-		
-		// 		// Use the fetched HTML data
-		// 		document.body.innerHTML = fetch_data;
-		
-		// 		//console.log(data);
-		// 		// Create a new script element
-		// 		var script = document.createElement('script');
-		
-		// 		// Set its src attribute to the extracted src
-		// 		script.src = scriptSrc;
-
-		// 		script.onload = function() {
-		// 			// This function will be called when the script is fully loaded and executed
-		// 			console.log('Script loaded');
-		// 			// Send WebSocket message here
-		// 			ws.send(JSON.stringify({type: 'tournament_lobby_update', 'tournament_id': server_data.tournament_id}));
-		// 		};
-				
-		// 		// Append the script element to the body of the document
-		// 		document.body.appendChild(script);
-		// 	})
-		// 	.catch((error) => {
-		// 		console.error('Error:', error);
-		// 	});
+		let tournamentGameUrl = '/tournament/' + server_data.tournament_id + '/';
+		history.pushState("", "", tournamentGameUrl);
+		router();
+		ws.send(JSON.stringify({type: 'tournament_lobby_update', 'tournament_id': server_data.tournament_id}));
+		// fetch_html_replace_dynamicDIV_activate_js('/tournament/' + server_data.tournament_id + '/', true, () => {
+		// 	this.ws.send(JSON.stringify({type: 'tournament_lobby_update', 'tournament_id': server_data.tournament_id}));
+		// });
+		console.log("join tournament");
 	}
+
+	
 }
 
-export default Lobby;
+
+
+customElements.define("pong-lobby", Lobby);
