+++ srcs/app_server/userManagement/static/userManagement/navbar.js
@@ -1,4 +1,4 @@
-import {fetch_html_replace_dynamicDIV_activate_js, fetch_with_internal_js, getCookie, executeJavaScriptInContent, ws} from "./land.js"
+import {ws} from "./land.js"
