+++ srcs/app_server/userManagement/static/userManagement/router.js
@@ -1,4 +1,4 @@
-import {fetch_html_replace_dynamicDIV_activate_js, fetch_with_internal_js, getCookie, executeJavaScriptInContent} from "./land.js"
+import {getCookie} from "./land.js"
 
 
 function router() {
